train_drhb/requirements.txt
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
numexpr=2.7.1=py37h423224d_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
pathlib2=2.3.5=py37_0
pillow=7.0.0=py37hb39fc2d_0
python=3.7.7=hcf32534_0_cpython
scikit-image=0.16.2=py37h0573a6f_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
sqlalchemy=1.3.16=py37h7b6447c_0
sqlite=3.31.1=h62c20be_1
statsmodels=0.11.0=py37h7b6447c_0
sympy=1.5.1=py37_0
tqdm=4.45.0=py_0
albumentations==0.4.5
apex==0.1
fastai==1.0.60
fastai2==0.0.17
fastcore==0.1.17
fastprogress==0.2.3
fastscript==0.1.4
opencv-python==4.2.0.34
openslide-python==1.1.1
optuna==1.3.0
python-editor==1.0.4
spacy==2.2.4
torch==1.5.0
torchvision==0.5.0
pandas==1.0.5

